ustream/requirements.txt
requests~=2.28.1
python-socketio~=5.7.1
aiohttp~=3.8.1